Create main.nf
Create the cutadapt module. 
This takes both single and paired end files and processing them accordingly. The read2 entry for the output tuple is optional, so remains empty if single end reads are used. This does not affect downstream processing (i.e. salmon)
An "info" file is created which contains the standard output. This contains information on how many times an adapter was trimmed, how many bases were quality clipped etc. and is required for multiqc