Update main.nf
Lots of changes:
1. Update module names in info
2. Added the instructions for cutadapt_options where user inputs their chosen options for the cutadapt function in a txt file
3. Added a section in the log info to read out the cutadapt options file
4. Import both new modules (FASTQC_CUTADAPT and CUTADAPT)
5. Added the cutadapt module
The first section deals with fastqc_cutadapt being passed without cutadapt.	fastqc_cutadapt and fastqc can't be run together if cutadapt is not also specified (they'd both received the same input and thus duplicate) so this produces an error message.
fastqc_cutadapt will work if cutadapt isn't specified, but the output of cutadapt (i.e. the trimmed files) is supplied in input. The directory MUST be named *cutadapt/ otherwise it will error
The second section deals with cutadapt being run. It imports the cutadapt options from the user provided txt file then runs cutadapt. It then asks if fastqc_cutadapt is also specified in modules_to_run and runs that if needed.
6. Removes the ch_salmon_multiqc section as this is not necessary now multiqc reads from the output directory. If salmon didn't run, there would be an empty channel which means multiqc couldn't be run on fastqc results alone in the original pipeline.
7. Remove the extra inputs to multiqc as they are no longer needed
8. Added a section to export the multiqc version used